requirements.txt
python-telegram-bot[all]
ollama
httpx
python-dotenv
PyYAML
SQLAlchemy
psycopg2
pytest
pytest-asyncio
pytest-cov
arabic-reshaper
python-bidi
numpy
debugpy